Give GitHubPane its own Package
Pre-initializing MEF is only required by GitHubPane.
